requirements.txt
Django==4.2
djangorestframework==3.16.1
django-filter==23.5
python-decouple==3.8
Pillow==12.1.1
psycopg2-binary==2.9.11
gunicorn==21.2.0
whitenoise==6.6.0
